backend/requirements.txt
# ============================================================
# VentureNode Backend Dependencies
# Latest stable versions as of March 2025
# ============================================================

# --- Core Framework ---
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20

# --- Configuration & Environment ---
pydantic>=2.10.0
pydantic-settings>=2.8.0
python-dotenv>=1.1.0

# --- Notion Integration ---
notion-client>=2.3.0

# --- LangGraph & LangChain Agent Orchestration ---
langgraph>=0.2.70
langchain>=0.3.20
langchain-core>=0.3.45
langchain-groq>=0.3.2
langchain-community>=0.3.19

# --- LLM Inference ---
groq>=0.20.0

# --- Market Research Tools (OSINT) ---
duckduckgo-search>=7.5.3
beautifulsoup4>=4.13.0
requests>=2.32.0
lxml>=5.3.0

# --- Vector Memory ---
faiss-cpu>=1.9.0
sentence-transformers>=3.4.0

# --- Rate Limiting & Middleware ---
slowapi>=0.1.9

# --- Observability & Logging ---
structlog>=25.1.0

# --- Authentication (Clerk JWT) ---
PyJWT[cryptography]>=2.9.0
